Rename document_store/protocols.py to document_store/protocol.py
